+++ lib/utopia/project.rb
@@ -1,13 +1,19 @@
 # frozen_string_literal: true
 
 # Released under the MIT License.
-# Copyright, 2020-2025, by Samuel Williams.
+# Copyright, 2020-2026, by Samuel Williams.
 
 require "utopia/project/version"
 
 require "variant"
 
+require "utopia/application"
+require "utopia/content"
+require "utopia/controller"
+require "utopia/exceptions"
 require "utopia/localization"
+require "utopia/redirection"
+require "utopia/static"
 
 require_relative "project/base"
 require_relative "project/import_map"
@@ -23,9 +29,9 @@ module Utopia
 		# The root directory for static assets.
 		PUBLIC_ROOT = File.expand_path("public", SITE_ROOT)
 		
-		# Appends a project application to the rack builder.
+		# Appends a project application to the protocol HTTP middleware builder.
 		#
-		# @parameter builder [Rack::Builder]
+		# @parameter builder [Protocol::HTTP::Middleware::Builder]
 		# @parameter root [String] The file-system root path of the project/gem.
 		# @parameter locales [Array(String)] an array of locales to support, e.g. `['en', 'ja']`.
 		def self.call(builder, root = Dir.pwd, locales: nil)
@@ -33,9 +39,6 @@ module Utopia
 				# Handle exceptions in production with a error page and send an email notification:
 				builder.use Utopia::Exceptions::Handler
 				builder.use Utopia::Exceptions::Mailer
-			else
-				# We want to propate exceptions up when running tests:
-				builder.use Rack::ShowExceptions unless UTOPIA.testing?
 			end
 			
 			# We serve static files from the project root:
@@ -46,7 +49,6 @@ module Utopia
 			builder.use Utopia::Redirection::Rewrite, {
 				"/" => "/index"
 			}
-			
 			builder.use Utopia::Redirection::DirectoryIndex
 			
 			builder.use Utopia::Redirection::Errors, {
@@ -67,7 +69,6 @@ module Utopia
 				# 'gallery' => Utopia::Gallery::Tags.new
 			}
 			
-			builder.run lambda {|env| [404, {}, []]}
 		end
 	end
 end