+++ template/config/environment.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+# Released under the MIT License.
+# Copyright, 2026, by Samuel Williams.
+
+require "utopia/setup"
+
+UTOPIA ||= Utopia.setup(File.expand_path("..", __dir__))